Change the path format of the HTTP controller to be device-focused.
This is a lot closer to the Python interface now we have c['deviceID'],
and removes the reliance on the many-methods-in-Controller approach.

Rather than individual methods being marked as @httpAccessible, it's now
a per-device option. Device code needs to be updated to pass this option
through to the Device superclass.