Don't require every remote-callable function to be `@expose`d.
Recent versions of Pyro made a breaking change, and by default any functions to
be called remotely must be marked with `@expose`.

Setting `REQUIRE_EXPOSE = false` reverts its behaviour to the previous.

Maybe it would be better to go through and mark methods individually on all the
different `Device`s, but expediency leaves that as an exercise for the
interested potential contributor.