Reconnect and reproxy devices if the controller goes away.
Since devices are given a random URI and port on startup, if the
controller is restarted then any cached proxies on clients will stop
working.

Here, we detect that scenario and refresh the proxy (based on its device
ID) so that any devices accessed with the `controller[]` method will
start working again after the controller restarts.